pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spritter"
version = "0.1.0"
description = "Core fuel price library"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "spritter" }]
dependencies = [
  "pillow>=10.0.0",
  "pytesseract>=0.3.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["spritter*"]
